Replace CREATE_NETWORK process
Replace with CREATE_BASE_NETWORK, TEST_PARAMETERS, THRESHOLD and CLUSTER

CREATE_BASE_NETWORK: Input is a directory name which contains sample and
count file. It converts the counts to TPM and aggregates to the gene level.
Then it creates and MCL correlation network with no threshold.
Outputs are the TPM file, tab file of GeneIDs to node ids, and the MCI file
The files are published to results as well

TEST_PARAMETERS: Inputs are the expt name and the TPMs file
This produces a basic network thresholded at 0.2 and varies the correlation
and k-nearest neighbour cutoffs and outputs network stats.
Output is the MCI file for the 0.2 network and stats on different correlation
and k-nearest neighbour cutoffs
The files are published to results as well

THRESHOLD: Inputs are the expt name and the MCI file from TEST_PARAMETERS
This uses the threshold and knn parameters to produce new thresholded networks
A max of 3 different networks are produced. Outputs are the new MCI files

CLUSTER: Inputs are the expt name and an MCI file. The process is run for
each value in the InflationParams list. It clusters the supplied MCI file
using the inflation value and outputs information about the clustering

New parameters have been added to nextflow.config