Add downloading of GO annotation file
GET_GO_ANNOTATION is included from the get_go_annotation module
In the workflow, it checks whether the GOFile already exists.
If it does, that path is passed to the GBA and ENRICHMENT processes.
If not, the GET_GO_ANNOTATION process is run and the output is the
path to the downloaded file which is published to the reference
directory.