Update GBA_SUMMARY
Change name to RUN_POST_GBA_STATS
Change to named outputs.
Also change the name of the output files so that they are unique.
Change from using the samples global parameter to passing it as an
input.
Change name of gba-analysis.R to post-gba-analysis.R
Had to make changes to run-GBA-network.R to deal with empty test files