Add filtering of genes based on zero values
bin/counts-to-fpkm-tpm.R
Added filter_threshold option to filter genes
If --filter_threshold is set the script calculates the distribution of
correlation values for genes split by the number of samples with zero values.
The filter_threshold is the number of abnormally high correlations (> 0.9) for
a given subset, above which those genes are removed. The script outputs a file
with 'filtered-by-zeros' in the filename with these genes removed.

main.nf
Changed CREATE_BASE_NETWORK to use the --filter_threshold option to
counts-to-fpkm-tpm.R

nextflow.config
Add high_cor_filter_threshold to config file